feature/ML-214/Cambios_ui-Recomendaciones (#148)
    Cambios de Labels.
    Cards ahora estan alineadas.
    Spinner para esperar a que las imagenes carguen.
    No podes clickear tus prendas cuando estas viendo recomendaciones viejas.
    Filtro ahora como Envudo.
    Se muestran las recomendaciones cercanas con el cosito de location.
    cambios menores en las estadisticas de promociones.

+++ web/src/app/anylitics/components/dashboard/dashboard.component.ts
@@ -247,7 +247,7 @@ export class DashboardComponent implements OnInit, OnDestroy {
     let sum = 0;
     for (let i = 0; i < filteredFeedBack.length; i++) {
       // tslint:disable-next-line: radix
-      sum = sum + parseInt(filteredFeedBack[i].feedBack);
+      sum = sum + parseFloat(filteredFeedBack[i].feedBack);
       console.log(`sumatoria parcial ` + sum);
     }
     this.feedBackProm = (sum / filteredFeedBack.length);
